新增帳戶相關 板面亂掉需要重新排板 尚未增加帳戶相關按鈕

+++ routes/web.php
@@ -20,8 +20,11 @@ Route::get('/', function () {
 
 Route::get('SurveyStatistics', [SurveyStatisticsController::class, 'index']);
 Route::get('SurveyStatistics/{id}', [SurveyStatisticsController::class, 'show'])->where('id', '[0-9]+')->name('Statistics.show');
-Route::get('SurveyStatistics/{id}/edit', [SurveyStatisticsController::class, 'edit'])->where('id', '[0-9]+')->name('Statistics.edit');
+Route::get('SurveyStatistics/{id}/edit', [SurveyStatisticsController::class, 'edit'])->where('id', '[0-9]+')->name('Statistics.edit')->middleware('can:admin-forest');
 Route::patch('SurveyStatistics/update/{id}', [SurveyStatisticsController::class, 'update'])->where('id', '[0-9]+')->name('Statistics.update');
-Route::delete('SurveyStatistics/delete/{id}', [SurveyStatisticsController::class, 'destroy'])->where('id', '[0-9]+')->name('Statistics.destroy');
-Route::get('SurveyStatistics/create', [SurveyStatisticsController::class, 'create'])->name('Statistics.create');
+Route::delete('SurveyStatistics/delete/{id}', [SurveyStatisticsController::class, 'destroy'])->where('id', '[0-9]+')->name('Statistics.destroy')->middleware('can:admin');
+Route::get('SurveyStatistics/create', [SurveyStatisticsController::class, 'create'])->name('Statistics.create')->middleware('can:admin');
 Route::post('SurveyStatistics/store', [SurveyStatisticsController::class, 'store'])->name('Statistics.store');
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
